requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow>=6.2.2
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==2.2.10

# Environment
django-environ==0.4.5

# Django REST fromework
djangorestframework>=3.11.0

# Passwords security
argon2-cffi==18.3.0

# Static files
whitenoise==4.1.2

# Celery
redis==3.4.1
django-redis==4.10.0
celery==4.4.2 
flower==0.9.2
tornado==4.2
